Split logic
